requirements.txt
# core
django == 1.11.2
django-oscar == 1.5
django-extensions == 1.9.7

# database
psycopg2 == 2.7.3.2

# haystack
Whoosh == 2.7.4
whitenoise == 3.3.1
pysolr == 3.6.0

django-environ == 0.4.4

# Please ensure that pillow, a fork of the the Python Imaging Library (PIL), gets installed with JPEG support.
# Supported formats are printed when pillow is first installed. Instructions on how to get JPEG support are highly
# platform specific, but guides for PIL should work for pillow as well. Generally speaking, you need to ensure that
# libjpeg-dev is installed and found during installation.
pillow == 4.3.0





# development
django-debug-toolbar == 1.8
pycodestyle
openpyxl == 2.4.9





# need for Heroku
gunicorn





# managements
# for oscar_populate_countries management
pycountry == 17.9.23

# utils
urllib3 == 1.22

# search
# Elasticsearch 6.x
# elasticsearch>=6.0.0,<7.0.0
solr == 0.4


